Create NeuralNetwork.java
A neural network consists of layers. The first layer is the input and the last layer is the output. The layers between the input and output are called hidden layers. The nodes of one layer can be connected to the nodes of the next layer.  Each connection between two nodes (called an edge) has a weight. 
And this is the java code that takes the number of layers, and nodes in each layer as input, and creates a class object.
 It sets each edge's weight from user input
And it can get the weights of any node in the layers
The program follows object oriented programming (OOPS)